Fix NumberFormatException by changing scratchRating and slopeRating from Int to Float
- Change scratchRating and slopeRating from Int? to Float? in RoundDetailDto.kt
- Change scratchRating and slopeRating from Int? to Float? in RoundDetail.kt
- Fixes NumberFormatException when API returns decimal values like 36.5
- All other models in codebase already use Float? correctly for these fields

+++ app/src/main/java/com/sogo/golf/msl/data/network/dto/mongodb/RoundDetailDto.kt
@@ -20,10 +20,10 @@ data class RoundDetailDto(
     val clubName: String? = null,
     
     @SerializedName("scratchRating")
-    val scratchRating: Int? = null,
+    val scratchRating: Float? = null,
     
     @SerializedName("slopeRating")
-    val slopeRating: Int? = null,
+    val slopeRating: Float? = null,
     
     @SerializedName("teeColor")
     val teeColor: String? = null,
